services/openclaw/requirements.txt
python-telegram-bot==20.7
groq==0.4.2
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
